Set up Devise for User Authentication
Add, install, and configure Devise features for user authentication. We want only the user owning the blog to be able to log in and add/edit posts, so comment out links to Sign Up and Log In. Sole user data will be added to database when app is started out