Add feature for user to get a suggestion blog post through Open AI Chat API when creating new blog post
The user clicks the button (i.e. submits the form in the 'new' view) which sends a POST request to /suggest_content.
The controller action is triggered, which uses OpenAiService to make a call to Open AI API. The action handles the response from the external API and a Turbo Stream format response is generated for the client.
This tells the client/browser to replace the content in the turbo frame with id="suggestion", populating the text area with the generated blog post suggestion.